### 2025-02-20 #7828 @editorialbot
[PRE REVIEW]: pycopm: An open-source tool to tailor OPM Flow geological models
**Submitting author:** @daavid00 (<a href="http://orcid.org/0000-0002-3343-1005">David Landa-Marbán</a>)
**Repository:** https://github.com/cssr-tools/pycopm
**Branch with paper.md** (empty if default branch):
**Version:** v2024.10
