+++ src/pycopm/core/pycopm.py
@@ -173,10 +173,16 @@ def load_parser():
         "-v",
         "--vicinity",
         default="",
-        help="The location to extract the sub model which can be assigned by a "
-        "region assignation, e.g., 'fipnum 2,4' extracts the cells with fipnums "
-        "equal to 2 or 4, or can be assigned by a polygon given the xy locations "
-        "in meters, e.g., 'xypolygon [0,0] [30,0] [30,30] [0,0]' ('' by default).",
+        help="The location to extract the sub model which can be assigned by "
+        "region values, e.g., 'fipnum 2,4' extracts the cells with fipnums "
+        "equal to 2 or 4, by a polygon given the xy locations in meters, e.g., "
+        "'xypolygon [0,0] [30,0] [30,30] [0,0]', or by the name of the well "
+        "and three different options for the neighbourhood: box, diamond, and diamondxy, where "
+        "for box the i, j, and k interval around the connections are given, e.g., 'welln box "
+        "[-1,1] [-2,2] [0,3]' results in a vicinity with 1 pm cell in the x direction, 2 pm "
+        "cells in the y direction and only 3 cells in the k positive direction, while the diamond "
+        "considers only the given number of cells around the well connections (e.g., 'welln "
+        "diamond 2') and diamondxy it is restricted to the xy plane ('' by default).",
     )
     parser.add_argument(
         "-c",
