+++ pyproject.toml
@@ -22,7 +22,7 @@ classifiers = [
     "Operating System :: OS Independent"
 ]
 dependencies = [
-    "ert",
+    "ert==13.0.4",
     "mako",
     "resdata"
 ]
