+++ src/pycopm/utils/parser_deck.py
@@ -126,10 +126,11 @@ def handle_schedulekw(dic, nrwo):
                     dic[name] = False
             if len(edit) > 1:
                 if edit[0][:2] != "--":
+                    well = edit[0].replace("'", "")
                     if name == "wsegvalv":
-                        if edit[0] in dic["swells"]:
+                        if well in dic["swells"]:
                             return True
-                    if edit[0] not in dic["nwells"]:
+                    if well not in dic["nwells"]:
                         return True
                 else:
                     return True
@@ -160,7 +161,8 @@ def get_wells_for_vicinity(dic):
                     if edit[0] == "/":
                         dic["edit0"] = ""
                         compdat = False
-                    if edit[0] not in [dic["optvic"], f"'{dic['optvic']}'"]:
+                    well = edit[0].replace("'", "")
+                    if well != dic["optvic"]:
                         continue
                     if len(edit) > 2:
                         if edit[0][:2] != "--":
@@ -194,12 +196,13 @@ def names_wells(dic):
                     if edit[0] == "/":
                         dic["edit0"] = ""
                         dic["compdat"] = False
-                    if edit[0] in dic["nwells"]:
+                    well = edit[0].replace("'", "")
+                    if well in dic["nwells"]:
                         continue
                 if len(edit) > 2:
                     if edit[0][:2] != "--":
-                        if edit[0] not in dic["awells"]:
-                            dic["awells"].append(edit[0])
+                        if well not in dic["awells"]:
+                            dic["awells"].append(well)
                         if (
                             dic["ic"][int(edit[1])]
                             * dic["jc"][int(edit[2])]
@@ -207,7 +210,7 @@ def names_wells(dic):
                             * dic["kc"][int(edit[4])]
                             > 0
                         ):
-                            dic["nwells"].append(edit[0])
+                            dic["nwells"].append(well)
             if nrwo == "COMPSEGS":
                 dic["compsegs"] = True
                 continue
@@ -218,8 +221,9 @@ def names_wells(dic):
                         dic["compsegs"] = False
                 if len(edit) > 1:
                     if edit[0][:2] != "--":
-                        if edit[0] in dic["awells"]:
-                            swell = edit[0]
+                        well = edit[0].replace("'", "")
+                        if well in dic["awells"]:
+                            swell = well
                         elif (
                             dic["ic"][int(edit[0])]
                             * dic["jc"][int(edit[1])]
@@ -241,13 +245,14 @@ def names_wells(dic):
         "wtest",
         "welopen",
         "wsegvalv",
+        "wecon",
     ]
     for name in dic["kw"]:
         dic[name] = False
     if dic["wvicinity"]:
         for n in ["s", "n", "a"]:
             dic[f"{n}wells"] = [
-                val
+                val.replace("'", "")
                 for val in dic[f"{n}wells"]
                 if val in [dic["optvic"], f"'{dic['optvic']}'"]
             ]
@@ -275,18 +280,20 @@ def names_segwells(dic):
                 if edit:
                     if edit[0] == "/":
                         dic["compdat"] = False
-                    if edit[0] in dic["swells"]:
+                    well = edit[0].replace("'", "")
+                    if well in dic["swells"]:
                         continue
                     if len(edit) > 2:
                         if edit[0][:2] != "--":
                             if dic["edit0"] == "":
                                 dic["edit0"] = nrwo.split()
                             else:
+                                well0 = dic["edit0"][0].replace("'", "")
                                 if (
[…]
