### 2025-04-07 #7828 @tberlok
> 👋 [@tberlok](https://github.com/tberlok) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This paper is outside my domain of expertise, which is computational astrophysics/plasma physics/hydrodynamics/cosmology, so I don't feel qualified to review the scientific merits of pycopm. But I would be willing to review installation, documentation and general programming aspects of the submission.
