### 2025-04-07 #7828 @gabersyd
James, sorry for the late answer, I have just seen your email. I could
review it but only for the end of April, as I am very busy now.

Gabriele

On Tue, Mar 25, 2025 at 9:11 PM James Gaboardi ***@***.***>
wrote:

> 👋 @gabersyd <https://github.com/gabersyd> - would you be willing to
> review this submission for JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
