### 2025-04-10 8bf04e6
Option to cell values instead of EQUIL in the SOLUTION
