+++ src/pycopm/utils/generate_files.py
@@ -58,20 +58,55 @@ def create_deck(dic):
     dic["flags1"] = (
         f"--parsing-strictness=low --output-mode=none --output-dir={dic['fol']}"
     )
+    types = [".INIT", ".EGRID"]
     if dic["mode"] in ["prep", "prep_deck", "all"]:
-        os.system(f"cp {dic['pth']}.DATA {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA")
-        print(
-            f"\nCloning {dic['pth']}.DATA to {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA for"
-            " the intial dry run to generate the grid (.EGRID) and static (.INIT) "
-            "properties\n"
-        )
-        os.system(
-            f"{dic['flow']} {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA {dic['flags']} "
-            f"--output-dir={dic['fol']}"
-        )
+        if dic["explicit"]:
+            types += [".UNRST"]
+            dic["lol"] = []
+            with open(dic["pth"] + ".DATA", "r", encoding="utf8") as file:
+                for row in csv.reader(file):
+                    nrwo = str(row)[2:-2].strip()
+                    if nrwo == "SCHEDULE":
+                        dic["lol"].append(nrwo)
+                        dic["lol"].append("RPTRST\n'BASIC=2'/\n")
+                        dic["lol"].append("TSTEP\n1*0.0001/\n")
+                        break
+                    dic["lol"].append(nrwo)
+            with open(
+                dic["deck"] + "_PREP_PYCOPM_DRYRUN.DATA",
+                "w",
+                encoding="utf8",
+            ) as file:
+                for row in dic["lol"]:
+                    file.write(row + "\n")
+            print(
+                f"\nTemporal {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA from {dic['pth']}.DATA for"
+                " the initial run to generate the grid (.EGRID), static (.INIT), and initial"
+                " (.UNRST) properties\n"
+            )
+            os.system(
+                f"{dic['flow']} {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA --output-mode=none "
+                f"--parsing-strictness=low --enable-opm-rst-file=1 --output-dir={dic['fol']}"
+            )
+            os.system(f"rm {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA")
+            os.system(f"cp {dic['pth']}.DATA {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA")
+            print(
+                f"\nCloning {dic['pth']}.DATA to {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA \n"
+            )
+        else:
+            os.system(f"cp {dic['pth']}.DATA {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA")
+            print(
+                f"\nCloning {dic['pth']}.DATA to {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA for"
+                " the initial dry run to generate the grid (.EGRID) and static"
+                " (.INIT) properties\n"
+            )
+            os.system(
+                f"{dic['flow']} {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA {dic['flags']} "
+                f"--output-dir={dic['fol']}"
+            )
         if dic["fol"] != os.path.abspath("."):
             os.system(f"mv {dic['deck']}_PREP_PYCOPM_DRYRUN.DATA " + f"{dic['fol']}")
-        for name in [".INIT", ".EGRID"]:
+        for name in types:
             files = f"{dic['fol']}/{dic['deck']}_PREP_PYCOPM_DRYRUN" + name
             if not os.path.isfile(files):
[…]
