### 2025-04-12 #7828 @jGaboardi
> Sadly I am not familiar with the reservoir modelling concepts so could not review the substance. I hope you find someone suitable [@jGaboardi](https://github.com/jGaboardi)

No problem at all. Thanks for the quick response.
