### 2025-04-27 #8115 @daavid00
Many thanks @gabersyd, and @tberlok for agreeing to review [pycopm](https://github.com/cssr-tools/pycopm), and @jGaboardi for agreeing to be the editor. I look forward to the reviewing process.
Best regards,
-David
