### 2025-05-02 cssr-tools/pycopm#42 @tberlok
Installing prerequisites on MacOS
Hello @daavid00 ,

I am following your instructions at https://cssr-tools.github.io/pycopm/installation.html for "Source build in MacOs"

I have installed boost, openblas and SuiteSparse using homebrew. I then store your instructions as a .sh script but it does not run successfully. Here is parts of the error message:

```
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:270:36: error: implicit instantiation of undefined template 'Dune::UMFPack<Dune::BCRSMatrix<Dune::FieldMatrix<double, 4, 4>>>'
        Dune::UMFPack<DiagMatWell> umfpackMatrix(duneD_, 0);
                                   ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:471:1: note: in instantiation of member function 'Opm::MultisegmentWellEquations<double, 4, 4>::extract' requested here
INSTANTIATE_TYPE(double)
^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:468:5: note: expanded from macro 'INSTANTIATE_TYPE'
    INSTANTIATE(T,4,4)      \
    ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:448:20: note: expanded from macro 'INSTANTIATE'
    template class MultisegmentWellEquations<T,numWellEq,numEq>;                               \
                   ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.hpp:36:25: note: template is declared here
template<class M> class UMFPack;
                        ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:270:36: error: implicit instantiation of undefined template 'Dune::UMFPack<Dune::BCRSMatrix<Dune::FieldMatrix<double, 4, 4>>>'
        Dune::UMFPack<DiagMatWell> umfpackMatrix(duneD_, 0);
                                   ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:471:1: note: in instantiation of member function 'Opm::MultisegmentWellEquations<double, 4, 5>::extract' requested here
INSTANTIATE_TYPE(double)
^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:469:5: note: expanded from macro 'INSTANTIATE_TYPE'
    INSTANTIATE(T,4,5)
    ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.cpp:448:20: note: expanded from macro 'INSTANTIATE'
    template class MultisegmentWellEquations<T,numWellEq,numEq>;                               \
                   ^
/Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellEquations.hpp:36:25: note: template is declared here
template<class M> class UMFPack;
                        ^
In file included from /Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellGeneric.cpp:33:
In file included from /Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/WellHelpers.hpp:26:
/Users/berlok/Dropbox/referee/joss-pycopm/dune-istl/dune/istl/bcrsmatrix.hh:1933:17: warning: declaration shadows a variable in namespace 'Dune::(anonymous)' [-Wshadow]
      real_type isNaN = 1;
                ^
/Users/berlok/Dropbox/referee/joss-pycopm/dune-common/dune/common/math.hh:323:27: note: previous declaration is here
    constexpr auto const &isNaN = Impl::MathDummy<MathImpl::isNaNImpl>::value;
                          ^
In file included from /Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/MultisegmentWellGeneric.cpp:33:
In file included from /Users/berlok/Dropbox/referee/joss-pycopm/opm-simulators/opm/simulators/wells/WellHelpers.hpp:26:
/Users/berlok/Dropbox/referee/joss-pycopm/dune-istl/dune/istl/bcrsmatrix.hh:1956:17: warning: declaration shadows a variable in namespace 'Dune::(anonymous)' [-Wshadow]
      real_type isNaN = 1;
                ^
/Users/berlok/Dropbox/referee/joss-pycopm/dune-common/dune/common/math.hh:323:27: note: previous declaration is here
    constexpr auto const &isNaN = Impl::MathDummy<MathImpl::isNaNImpl>::value;
                          ^
20 warnings and 9 errors generated.
make[3]: *** [CMakeFiles/opmsimulators.dir/opm/simulators/wells/MultisegmentWellEquations.cpp.o] Error 1
make[3]: *** Waiting for unfinished jobs....
4 warnings generated.
18 warnings generated.
20 warnings generated.
20 warnings generated.
make[2]: *** [CMakeFiles/opmsimulators.dir/all] Error 2
make[1]: *** [CMakeFiles/flow.dir/rule] Error 2
make: *** [flow] Error 2

```
