### 2025-05-02 cssr-tools/pycopm#42 @daavid00
Hi @tberlok,

I am glad to know you also use macOS. I have run the script in a new folder and I manage to install OPM Flow. These are the specifications I used (I tried to frequently upgrade to the latest available versions): boost 1.88.0, openblas 0.3.29, suite-sparse 7.10.2, cmake 4.0.1, Apple clang version 17.0.0 (clang-1700.0.13.3).
I can take a look at the whole build process to try to find the issue, then please rerun the script and attach the generated output (e.g., source build_opm_mpi.sh > output.txt).
