### 2025-05-05 cssr-tools/pycopm#42 @tberlok
Thanks for the quick reply. I had the same versions as you except for clang, where I was running version 15. I upgraded my OS and the command line tools and it now also works for me.
