### 2025-05-05 cssr-tools/pycopm#42 @tberlok
Based on my issues with installing I suggest adding the following instructions to the build from source part:

Install dependencies with homebrew:

brew install boost
brew install openblas
brew install cmake
brew install suite-sparse

Update your clang compiler to version 17.

Create a virtual environment using Python 3.11 or above:

```
python3.11 -m venv vpycopm
```

```
brew install python@3.11
```

(Current instructions)

```
deactivate
source vpycopm/bin/activate
```

Check that you can import opm
```
python -c "import opm"
```
Check that you have flow in your path
```
flow --help
```

If not, add `build/opm-simulators/bin` to your path.
