### 2025-05-05 cssr-tools/pycopm#42 @tberlok
For running the example, can you change the instruction to something like:

```
cd pycopm/examples/configurations/drogon
pycopm -i input.toml -o drogon_coarser
```

When I run this first example, I get:

```
Missmatch (standard simulation from opm-test deck):  6.0387e+00
Missmatch (single simulation):  9.8843e+00
Difference (webviz - pycopm): -34.42 (a positive number (percentage) is the goal)
```

which I guess is not good given that the number is negative?

And I also still have trouble running all the tests. I get the following errors:

```
The initial/dry run of MODEL3.DATA succeeded

Initializing pycopm to generate the coarsened files, please wait
-------------------------------------------------------------------- Captured stderr call --------------------------------------------------------------------
Traceback (most recent call last):
  File "/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/vpycopm/bin/pycopm", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/vpycopm/lib/python3.11/site-packages/pycopm/core/pycopm.py", line 445, in main
    pycopm()
  File "/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/vpycopm/lib/python3.11/site-packages/pycopm/core/pycopm.py", line 92, in pycopm
    create_deck(dic)
  File "/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/vpycopm/lib/python3.11/site-packages/pycopm/utils/generate_files.py", line 164, in create_deck
    initialize_variables(dic)
  File "/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/vpycopm/lib/python3.11/site-packages/pycopm/utils/generate_files.py", line 559, in initialize_variables
    dic["ogrid"] = OpmFile(dic["deck"] + ".EGRID")
                   ^^^^^^^
NameError: name 'OpmFile' is not defined
================================================================== short test summary info ===================================================================
FAILED tests/test_2_coarsening.py::test_coarsening - subprocess.CalledProcessError: Command '['pycopm', '-i', '/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/pycopm/examples/decks/HELLO_WORL...
FAILED tests/test_3_refinement.py::test_refinement - subprocess.CalledProcessError: Command '['pycopm', '-i', '/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/pycopm/examples/decks/MODEL2.DAT...
FAILED tests/test_4_submodel.py::test_submodel - subprocess.CalledProcessError: Command '['pycopm', '-i', '/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/pycopm/examples/decks/MODEL0.DAT...
FAILED tests/test_5_transform.py::test_transform - subprocess.CalledProcessError: Command '['pycopm', '-i', '/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/pycopm/examples/decks/MODEL0.DAT...
FAILED tests/test_6_segwell_faults.py::test_complex - subprocess.CalledProcessError: Command '['pycopm', '-o', '/Users/berlok/Library/CloudStorage/Dropbox/referee/joss-pycopm/pycopm/tests/output/complex', '-...
========================================================== 5 failed, 2 passed in 181.47s (0:03:01) ===========================================================
(vpycopm)
(base) ~/.../joss-pycopm/pycopm [main]

```
