+++ examples/configurations/drogon/hm.toml
@@ -9,7 +9,7 @@ Y = [0,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0
 Z = [0,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0] # Array of z-coarsening
 net = 50 # Number of ensembles
 mep = 2 # Maximum number of ensembles running in parallel
-mrt = 120 # Maximum runtime in seconds of a realization. A value of 0 means unlimited runtime. 
+mrt = 300 # Maximum runtime in seconds of a realization. A value of 0 means unlimited runtime. 
 mrn = 2 # Minimum number of realizations that must have succeeded for the simulation to be regarded as a success.
 rds = 7 # Set a specific seed for reproducibility. A value of 0 means no seed.  
 obs = "observations_training" # Name of the observation file for the hm ('observations_training', 'observations_test', or 'observations_complete')

[…]
