+++ docs/text/pycopm.utils.runs_executer.rst
@@ -3,6 +3,6 @@ pycopm.utils.runs\_executer module
 
 .. automodule:: pycopm.utils.runs_executer
    :members:
-   :undoc-members:
-   :show-inheritance:
    :private-members:
+   :show-inheritance:
+   :undoc-members:
