+++ docs/_sources/pycopm.utils.generate_files.rst.txt
@@ -3,6 +3,6 @@ pycopm.utils.generate\_files module
 
 .. automodule:: pycopm.utils.generate_files
    :members:
-   :undoc-members:
-   :show-inheritance:
    :private-members:
+   :show-inheritance:
+   :undoc-members:
