+++ docs/_sources/pycopm.utils.mapping_methods.rst.txt
@@ -3,6 +3,6 @@ pycopm.utils.mapping\_methods module
 
 .. automodule:: pycopm.utils.mapping_methods
    :members:
-   :undoc-members:
-   :show-inheritance:
    :private-members:
+   :show-inheritance:
+   :undoc-members:
