+++ docs/_sources/pycopm.utils.parser_deck.rst.txt
@@ -3,6 +3,6 @@ pycopm.utils.parser\_deck module
 
 .. automodule:: pycopm.utils.parser_deck
    :members:
-   :undoc-members:
-   :show-inheritance:
    :private-members:
+   :show-inheritance:
+   :undoc-members:
