+++ pyproject.toml
@@ -19,9 +19,11 @@ classifiers = [
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Operating System :: OS Independent"
 ]
 dependencies = [
+    "alive_progress",
     "ert==13.0.4",
     "mako",
     "resdata"
