### 2025-06-02 cssr-tools/pycopm#42 @tberlok
Hi @daavid00 ,

Sorry for taking this long to get back to you. I am still not able to load OpmFile. I don't think it is the path, because I can load opm.

```
Cell In[5], line 1
----> 1 from opm.io.ecl import EclFile as OpmFile

File ~/Dropbox/referee/joss-pycopm/build/opm-common/python/opm/io/__init__.py:1
----> 1 from .parser import *
      2 from .log import *
      3 from opm._common import UnitSystem

File ~/Dropbox/referee/joss-pycopm/build/opm-common/python/opm/io/parser/__init__.py:1
----> 1 from opm._common import action
      2 from opm._common import Parser
      3 from opm._common import ParseContext

File ~/Dropbox/referee/joss-pycopm/build/opm-common/python/opm/_common.py:12
      1 # This is the entry point were all the pybind11/C++ symbols are imported into
      2 # Python. Before actually being used the symbols are typically imported one
      3 # more time to a more suitable location; e.g the Parser() class is imported in
   (...)      9 #
     10 #   from opm.io.parser import Parser
     11 from __future__ import absolute_import
---> 12 from .opmcommon_python import action
     14 from .opmcommon_python import Parser, ParseContext, Builtin, eclSectionType
     15 from .opmcommon_python import DeckKeyword

ModuleNotFoundError: No module named 'opm.opmcommon_python'
```
