### 2025-06-02 cssr-tools/pycopm#42 @daavid00
Hi @tberlok,

No problem, and thanks again for your time reviewing this repository.

Getting to work the Python opm library in macOS it is not straight forward, that is why this is now an optional dependency, including the running of the tests (after https://github.com/cssr-tools/pycopm/pull/43). Then one can use all implemented functionality in pycopm using the default dependency resdata.
