### 2025-06-20 69b9975
Writing map to global index for submodels
