+++ src/pycopm/utils/mapping_methods.py
@@ -356,7 +356,7 @@ def add_pv_bc(dic):
                 )
                 pvy += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["opernum_c"][ins] = "2"
+                dic["opernum_c"][ins] = f"{ind+dic['xn']}"
                 nums += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvy)
@@ -374,7 +374,7 @@ def add_pv_bc(dic):
                     )
                     pvy += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["opernum_c"][ins] = "2"
+                        dic["opernum_c"][ins] = f"{ind}"
                         ijs[i] = j + 1
                         nums += 1.0
                         if dic["pvcorr"] in [1, 2]:
@@ -401,7 +401,7 @@ def add_pv_bc(dic):
                     inb = ind - dic["xn"]
                 pvy += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["opernum_c"][ins] = "2"
+                dic["opernum_c"][ins] = f"{inb}"
                 numn += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvy)
@@ -422,7 +422,7 @@ def add_pv_bc(dic):
                     )
                     pvy += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["opernum_c"][ins] = "2"
+                        dic["opernum_c"][ins] = f"{ind}"
                         ijn[i] = j + 1
                         numn += 1.0
                         if dic["pvcorr"] in [1, 2]:
@@ -446,7 +446,7 @@ def add_pv_bc(dic):
                 )
                 pvx += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["opernum_c"][ins] = "2"
+                dic["opernum_c"][ins] = f"{ind+1}"
                 nume += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvx)
@@ -464,7 +464,7 @@ def add_pv_bc(dic):
                     )
                     pvx += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["opernum_c"][ins] = "2"
+                        dic["opernum_c"][ins] = f"{ind+1}"
                         ije[j] = i + 1
                         nume += 1.0
                         if dic["pvcorr"] in [1, 2]:
@@ -491,7 +491,7 @@ def add_pv_bc(dic):
                     inb = ind - 1
                 pvx += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["opernum_c"][ins] = "2"
+                dic["opernum_c"][ins] = f"{inb}"
                 numw += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvx)
@@ -514,7 +514,7 @@ def add_pv_bc(dic):
                     )
                     pvx += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["opernum_c"][ins] = "2"
+                        dic["opernum_c"][ins] = f"{ind-1}"
                         ijw[j] = i + 1
[…]
