### 2025-06-20 bceaf21
Renaming to SUBTOGLOB for the mapping of indices
