+++ src/pycopm/utils/generate_files.py
@@ -780,49 +780,52 @@ def handle_nnc_trans(dic):
             indel.append(i)
     rnnc1 = np.delete(rnnc1, indel)
     rnnc2 = np.delete(rnnc2, indel)
-    for n, (n1, n2) in enumerate(zip(cnnc1, cnnc2)):
-        if dic["resdata"]:
-            ijk1 = coag.get_ijk(global_index=n1 - 1)
-            ijk2 = coag.get_ijk(global_index=n2 - 1)
-        else:
-            ijk1 = coag.ijk_from_global_index(n1 - 1)
-            ijk2 = coag.ijk_from_global_index(n2 - 1)
-        fip1 = ijk1[2] + 1
-        fip2 = ijk2[2] + 1
-        rtran = 0
-        found = 0
-        indel = []
-        for i, (r1, r2) in enumerate(zip(rnnc1, rnnc2)):
+    print("Processing the transmissibilities")
+    with alive_bar(len(cnnc1)) as bar_animation:
+        for n, (n1, n2) in enumerate(zip(cnnc1, cnnc2)):
+            bar_animation()
             if dic["resdata"]:
-                rijk1 = dic["grid"].get_ijk(global_index=r1 - 1)
-                rijk2 = dic["grid"].get_ijk(global_index=r2 - 1)
+                ijk1 = coag.get_ijk(global_index=n1 - 1)
+                ijk2 = coag.get_ijk(global_index=n2 - 1)
             else:
-                rijk1 = dic["grid"].ijk_from_global_index(r1 - 1)
-                rijk2 = dic["grid"].ijk_from_global_index(r2 - 1)
-            rfip1 = dic["kc"][rijk1[2] + 1]
-            rfip2 = dic["kc"][rijk2[2] + 1]
-            if (
-                ijk1[0] == rijk1[0]
-                and ijk2[0] == rijk2[0]
-                and ijk1[1] == rijk1[1]
-                and ijk2[1] == rijk2[1]
-                and fip1 == rfip1
-                and fip2 == rfip2
-                and fip1 != fip2
-                and (ijk1[0] != ijk2[0] or ijk1[1] != ijk2[1])
-            ):
-                rtran += rnnct[i]
-                found = 1
-                indel.append(i)
-        rnnc1 = np.delete(rnnc1, indel)
-        rnnc2 = np.delete(rnnc2, indel)
-        if found == 1:
-            mult = rtran / cnnct[n]
-        else:
-            mult = 0
-        dic["coa_editnnc"].append(
-            f"{ijk1[0]+1} {ijk1[1]+1} {ijk1[2]+1} {ijk2[0]+1} {ijk2[1]+1} {ijk2[2]+1} {mult} /"
-        )
+                ijk1 = coag.ijk_from_global_index(n1 - 1)
+                ijk2 = coag.ijk_from_global_index(n2 - 1)
+            fip1 = ijk1[2] + 1
+            fip2 = ijk2[2] + 1
+            rtran = 0
+            found = 0
+            indel = []
+            for i, (r1, r2) in enumerate(zip(rnnc1, rnnc2)):
+                if dic["resdata"]:
+                    rijk1 = dic["grid"].get_ijk(global_index=r1 - 1)
+                    rijk2 = dic["grid"].get_ijk(global_index=r2 - 1)
+                else:
+                    rijk1 = dic["grid"].ijk_from_global_index(r1 - 1)
+                    rijk2 = dic["grid"].ijk_from_global_index(r2 - 1)
+                rfip1 = dic["kc"][rijk1[2] + 1]
+                rfip2 = dic["kc"][rijk2[2] + 1]
+                if (
+                    ijk1[0] == rijk1[0]
+                    and ijk2[0] == rijk2[0]
+                    and ijk1[1] == rijk1[1]
[…]
