### 2025-06-25 d9d17a3
Improvements to submodels
