+++ src/pycopm/utils/mapping_methods.py
@@ -288,6 +288,18 @@ def add_pv_bc(dic):
         dic (dict): Modified global dictionary
 
     """
+    if (
+        dic["hvicinity"] in ["diamond", "diamondxy"]
+        and int(dic["vicinity"].split(" ")[2]) > 0
+    ):
+        for well in dic["wvicinity"]:
+            dic["subgind"].append(
+                well[0]
+                - dic["mini"]
+                + (well[1] - dic["minj"] + 1) * dic["nx"]
+                + (well[2] - dic["mink"] + 1) * dic["nx"] * dic["ny"]
+                + 1
+            )
     nbct = 0
     for k in range(dic["nz"]):
         k_r = k + dic["mink"] - 1
@@ -356,7 +368,8 @@ def add_pv_bc(dic):
                 )
                 pvy += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["subtoglob_c"][ins] = f"{ind+dic['xn']}"
+                if ins not in dic["subgind"]:
+                    dic["subtoglob_c"][ins] = f"{ind+dic['xn']}"
                 nums += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvy)
@@ -374,7 +387,8 @@ def add_pv_bc(dic):
                     )
                     pvy += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["subtoglob_c"][ins] = f"{ind}"
+                        if ins not in dic["subgind"]:
+                            dic["subtoglob_c"][ins] = f"{ind}"
                         ijs[i] = j + 1
                         nums += 1.0
                         if dic["pvcorr"] in [1, 2]:
@@ -401,7 +415,8 @@ def add_pv_bc(dic):
                     inb = ind - dic["xn"]
                 pvy += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["subtoglob_c"][ins] = f"{inb}"
+                if j_n == 0 and ins not in dic["subgind"]:
+                    dic["subtoglob_c"][ins] = f"{inb}"
                 numn += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvy)
@@ -422,7 +437,8 @@ def add_pv_bc(dic):
                     )
                     pvy += 0.5 * dic["porv"][ind]
                     if int(dic["actnum_c"][ins]) > 0:
-                        dic["subtoglob_c"][ins] = f"{ind}"
+                        if ins not in dic["subgind"]:
+                            dic["subtoglob_c"][ins] = f"{ind-dic['xn']}"
                         ijn[i] = j + 1
                         numn += 1.0
                         if dic["pvcorr"] in [1, 2]:
@@ -446,35 +462,44 @@ def add_pv_bc(dic):
                 )
                 pvx += dic["porv"][ind]
             if int(dic["actnum_c"][ins]) > 0:
-                dic["subtoglob_c"][ins] = f"{ind+1}"
+                if ins not in dic["subgind"]:
+                    dic["subtoglob_c"][ins] = f"{ind+1}"
                 nume += 1.0
                 if dic["pvcorr"] in [1, 2]:
                     dic["porv_c"][ins] = str(float(dic["porv_c"][ins]) + pvx)
                 else:
                     pve += pvx
-            elif dic["porv"][ind + 1] > 0:
-                for i in range(dic["nx"] - 1 - i_0):
-                    ins = (
[…]
