### 2025-06-30 487baed
Compact format to write values and improving performance
