+++ src/pycopm/utils/generate_files.py
@@ -734,8 +734,9 @@ def handle_nnc_trans(dic):
         coa_dz[coapv > 0] = np.array(coa["DZ"])
         ref_dz[refpv > 0] = np.array(dic["ini"]["DZ"])
     dic["coa_editnnc"] = []
+    dic["nnct"] = [[[] for _ in range(dic["yn"])] for _ in range(dic["xn"])]
     indel = []
-    for i, (r1, r2) in enumerate(zip(rnnc1, rnnc2)):
+    for r1, r2, trn in zip(rnnc1, rnnc2, rnnct):
         if dic["resdata"]:
             rijk1 = dic["grid"].get_ijk(global_index=r1 - 1)
             rijk2 = dic["grid"].get_ijk(global_index=r2 - 1)
@@ -752,7 +753,7 @@ def handle_nnc_trans(dic):
                     + (dic["kc"][rijk1[2] + 1] - 1) * dic["nx"] * dic["ny"]
                 )
                 if coa_dz[ind] > 0:
-                    dic["coa_tranx"][ind] += rnnct[i] * ref_dz[r1 - 1] / coa_dz[ind]
+                    dic["coa_tranx"][ind] += trn * ref_dz[r1 - 1] / coa_dz[ind]
             elif rijk1[1] + 1 == rijk2[1]:
                 ind = (
                     (dic["ic"][rijk1[0] + 1] - 1)
@@ -760,7 +761,7 @@ def handle_nnc_trans(dic):
                     + (dic["kc"][rijk1[2] + 1] - 1) * dic["nx"] * dic["ny"]
                 )
                 if coa_dz[ind] > 0:
-                    dic["coa_trany"][ind] += rnnct[i] * ref_dz[r1 - 1] / coa_dz[ind]
+                    dic["coa_trany"][ind] += trn * ref_dz[r1 - 1] / coa_dz[ind]
             elif rijk1[0] == rijk2[0] + 1:
                 ind = (
                     (dic["ic"][rijk2[0] + 1] - 1)
@@ -768,7 +769,7 @@ def handle_nnc_trans(dic):
                     + (dic["kc"][rijk2[2] + 1] - 1) * dic["nx"] * dic["ny"]
                 )
                 if coa_dz[ind] > 0:
-                    dic["coa_tranx"][ind] += rnnct[i] * ref_dz[r2 - 1] / coa_dz[ind]
+                    dic["coa_tranx"][ind] += trn * ref_dz[r2 - 1] / coa_dz[ind]
             elif rijk1[1] == rijk2[1] + 1:
                 ind = (
                     (dic["ic"][rijk2[0] + 1] - 1)
@@ -776,10 +777,9 @@ def handle_nnc_trans(dic):
                     + (dic["kc"][rijk2[2] + 1] - 1) * dic["nx"] * dic["ny"]
                 )
                 if coa_dz[ind] > 0:
-                    dic["coa_trany"][ind] += rnnct[i] * ref_dz[r2 - 1] / coa_dz[ind]
-            indel.append(i)
-    rnnc1 = np.delete(rnnc1, indel)
-    rnnc2 = np.delete(rnnc2, indel)
+                    dic["coa_trany"][ind] += trn * ref_dz[r2 - 1] / coa_dz[ind]
+        else:
+            dic["nnct"][rijk1[0]][rijk1[1]].append([rijk1[2], rijk2[2], trn])
     print("Processing the transmissibilities")
     with alive_bar(len(cnnc1)) as bar_animation:
         for n, (n1, n2) in enumerate(zip(cnnc1, cnnc2)):
@@ -790,42 +790,26 @@ def handle_nnc_trans(dic):
             else:
[…]
