### 2025-07-16 #8115 @gabersyd
I encountered so many issues installing OPM on my OSX system, that I
stopped trying. It first required me to install BOOST, then dune-common,
and I don't know what comes next. Somehow either a mac friendly
installation of OPM is provided in pycopm or the software can be made
independent from OPM. Whatever is easier.

I will try to install it on LONI, the Louisiana Supercomputer system, which
is just a linux system. If that is simpler, I will review the linux
installation.

On Wed, Jul 16, 2025 at 2:56 PM James Gaboardi ***@***.***>
wrote:

> *jGaboardi* left a comment (openjournals/joss-reviews#8115)
> <https://github.com/openjournals/joss-reviews/issues/8115#issuecomment-3078463255>
>
> I am working on it now. Testing its installation on my OSX laptop is not
> straightforward. One needs to install the OPM database first but binaries
> for OPM are available only in Windows and Linux (
> https://opm-project.org/?page_id=36). I am installing it from the source.
> Not that straightforward. I would consider adding an automatic OSX
> installation of OPM in pycopm since it is so important.
>
> Gabriele
> … <#m_-1004746733041186354_>
>
> @gabersyd <https://github.com/gabersyd> This is a constructive
> observation. It looks like @tberlok <https://github.com/tberlok> has
> already opened an issue for that over in the repo itself –
> cssr-tools/pycopm#42 <https://github.com/cssr-tools/pycopm/issues/42>.
> Let's continue the conversation over there.
>
