### 2025-07-16 #8115 @jGaboardi
@gabersyd It would be better for these conversations to happen in the `pycopm` repo itself. Please either continue on the thread in https://github.com/cssr-tools/pycopm/issues/42 or open a fresh issue detailing the failures for @daavid00 to address.
