### 2025-07-17 6471252
Message for missing optional OPM python when running tests
