+++ tests/test_4_submodel.py
@@ -7,6 +7,7 @@
 import os
 import pathlib
 import subprocess
+import warnings
 import numpy as np
 from resdata.resfile import ResdataFile
 
@@ -14,7 +15,13 @@ OPM = False
 try:
     OPM = bool(__import__("opm"))
 except ImportError:
-    pass
+    warnings.warn(
+        UserWarning(
+            "The optional OPM Python package for postprocessing "
+            "the simulation files is not found; then, test_4_submodel.py is only run "
+            "using resdata."
+        )
+    )
 
 testpth: pathlib.Path = pathlib.Path(__file__).parent
 mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
