+++ tests/test_2_coarsening.py
@@ -7,12 +7,19 @@
 import os
 import pathlib
 import subprocess
+import warnings
 
 OPM = False
 try:
     OPM = bool(__import__("opm"))
 except ImportError:
-    pass
+    warnings.warn(
+        UserWarning(
+            "The optional OPM Python package for postprocessing "
+            "the simulation files is not found; then, test_2_coarsening.py is only run "
+            "using resdata."
+        )
+    )
 
 testpth: pathlib.Path = pathlib.Path(__file__).parent
 mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
