### 2025-07-17 cssr-tools/pycopm#42 @daavid00
> I encountered so many issues installing OPM on my OSX system, that I
> stopped trying. It first required me to install BOOST, then dune-common,
> and I don't know what comes next. Somehow either a mac friendly
> installation of OPM is provided in pycopm or the software can be made
> independent from OPM. Whatever is easier.

Hi @gabersyd, thanks again for your time to review this. The OPM Flow simulator is a core component of pycopm. As you mentioned, unfortunately installing OPM Flow in macOS is not straight forward, i.e., no binaries and it needs to be build from source with a few dependencies. An attempt to ease this has been done by adding documentation:

https://cssr-tools.github.io/pycopm/installation.html#source-build-in-macos

I can take a look at the whole build process to try to find the issue, to this end you could save the commands in the source-build-in-macos documentation (building dune + OPM Flow) and share the terminal output  (e.g., source build_opm_macOS.sh > output.txt).
