+++ src/pycopm/utils/generate_files.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2024 NORCE
+# SPDX-FileCopyrightText: 2024-2025 NORCE Research AS
 # SPDX-License-Identifier: GPL-3.0
 # pylint: disable=R0912,R0913,R0914,R0915,C0302,R0917,R1702,R0916,R0911,R0801,E1102
 
@@ -14,7 +14,6 @@ import numpy as np
 from alive_progress import alive_bar
 from resdata.grid import Grid
 from resdata.resfile import ResdataFile
-from mako.template import Template
 from pycopm.utils.parser_deck import process_the_deck
 from pycopm.utils.mapping_methods import (
     add_pv_bc,
@@ -239,10 +238,10 @@ def create_deck(dic):
                                             + k * dic["xn"] * dic["yn"]
                                         )
                                         for _ in range(dic["X"][i] + 1):
-                                            dic[f"{name}_c"][n] = str(
-                                                int(dic[name][ind])
+                                            dic[f"{name}_c"][n] = (
+                                                f"{int(dic[name][ind])}"
                                                 if "num" in name
-                                                else dic[name][ind]
+                                                else f"{dic[name][ind]:E}"
                                             )
                                             n += 1
             dic["actnum_c"] = ["1" if float(val) > 0 else "0" for val in dic["porv_c"]]
@@ -429,14 +428,15 @@ def create_deck(dic):
             )
             cor_pa[cor_fipo > thr] *= 1 + fact
             cor_pv[cor_pv > 0] = cor_pa
+            cor_pv = compact_format("".join(f"{val} " for val in cor_pv).split())
             with open(
                 f"{dic['fol']}/{dic['label']}PORV.INC",
                 "w",
                 encoding="utf8",
             ) as file:
                 file.write("PORV\n")
-                file.write("\n".join(f"{val}" for val in cor_pv))
-                file.write("\n/")
+                file.write("".join(cor_pv))
+                file.write("/")
             os.system(
                 f"{dic['flow']} {dic['fol']}/{dic['write']}_CORR.DATA {dic['flags1']}"
             )
@@ -461,14 +461,15 @@ def create_deck(dic):
             )
             cor_pa[cor_sgas > thr] *= 1 + fact
             cor_pv[cor_pv > 0] = cor_pa
+            cor_pv = compact_format("".join(f"{val} " for val in cor_pv).split())
             with open(
                 f"{dic['fol']}/{dic['label']}PORV.INC",
                 "w",
                 encoding="utf8",
             ) as file:
                 file.write("PORV\n")
-                file.write("\n".join(f"{val}" for val in cor_pv))
-                file.write("\n/")
+                file.write("".join(cor_pv))
+                file.write("/")
             print(
                 f"\nRunning {dic['fol']}/{dic['write']}_CORR.DATA with the corrected "
                 "pore volume\n"
@@ -493,10 +494,6 @@ def create_deck(dic):
         subprocess.run(
             [dic["flow"], f"{dic['write']}.DATA"] + dic["flags"].split(" "), check=False
         )
-        # It seems there is a bug for dryruns in flow, commenting the following until fixing this
-        # prosc = subprocess.run(
-        #     [dic["flow"], f"{dic['write']}.DATA"] + dic["flags"].split(" "), check=False
-        # )
[…]
