### 2025-08-13 7b09eb8
Adding another paper using pycopm and small fix in terminal message
