### 2025-10-02 #8115 @jGaboardi
@daavid00

At this point I'll need to recommend a reject to the Track Editor-in-Chief as the best way forward. However, your manuscript is sound and there is clearly substantial scholarly effort. As noted by the reviewers (which I have verified myself) the parts that are really missing are the documentation, and by extension reproducibility and contributing (as also seen in the reviewers' checklists). I did verify that installation on macOS was now successful after your updated instructions, but I was not able to perform any examples or get `pytest` to run successfully. Therefore, I would encourage you to resubmit after spending more time on these things. Perhaps even have a colleague or two run through installation, set up, etc. to verify the steps are both (a) clear for a 3rd party; and (b) functional on various platforms. To be 100% transparent and clear – my comments here should not be considered as either exhaustive or a definitive checklist that will guarantee acceptance upon resubmission.
