### 2025-10-03 #8115 @jGaboardi
@daavid00

I don't have the error I was getting from `pytest` at hand, and unfortunately don't have the bandwidth to recreate everything. However, I do remember that `pytest` was able to run but that all tests failed. Hope this helps.
